refactor(eval): extract shared.rs and anthropic.rs modules
Move shared eval infrastructure (BPE tokenizer, embedder, entity
extraction helper) into eval::shared and Anthropic API/Batch client
functions into eval::anthropic. Backward-compat re-exports keep all
existing call sites in token_efficiency.rs, longmemeval.rs, and
locomo.rs working without changes.
